added toolbar to cardview
